Initial commit of the uwsgi spool editor
This script can be used to print the contents of a uwsgi spool file, or
to edit the value side of a key in the uwsgi spool file. The modified
file is written back out with the name of the original file and
'.update' appended to it.

The original spool file is not modified in anyway, Most likely your
looking for a destructive change in which case you should delete the
original file once your done, assuming the modification succeed.

The script can also be used to manually add key=value pairs to the spool
file, though it does not do any size checking or validation.