Rewrite spool editor to be more correct
Previous attempt at this was I believe capable of writing out garbage
files after updates, It also didn't know how to pass the special 'body'
spooler key behavour.

This commit adds breaks each section of the task/spooler down into a
separate class that is able to serialize & deserialize itself [except
body, which is just arbitrary data].

It also cleans up a few of the functions and adds doc strings and type
hints (because this is 2021).

Known issues:
- I haven't tested this with spoolers with 'at' times [special execution
  time frames].
- I haven't tested this with spoolers that have a priority
- This isn't checking any of the file locking, which it probably should.